Troubleshooting & Testing
We were getting a weird gradle error. So, the update was probably the source of the issue. I ran some troubleshooting stuff, which I think helped somewhat. However, probably the biggest problem was actually securely.  It was blocking a download of a certain gradle file. Note to self, avoid on campus networks for compilation. Otherwise, we added functionality to the teleop program.